Added publication management, minor fixes
Added support for publication management in add and remove functionality. Edit functionality was deemed not immediately necessary. Publications are uploaded to atlas-769c1's cloud storage and stored there, linked to records about the publication in the database by filename consistency (i.e. the name in storage is the same as the 'name' variable in the database). For this reason, uploading two files by the same name is unacceptable.

Minor changes were made to other functions as well. Checkboxes were removed from the entities, some ids were changed, and a better-looking CSS is starting to be implemented.